+++ src/backend/controllers/utils/interfaces/corpusSuggestionInterfaces.ts
@@ -1,0 +1,4 @@
+import { Corpus } from './corpusInterfaces';
+import { Suggestion } from './suggestionInterfaces';
+
+export interface CorpusSuggestion extends Corpus, Suggestion {}
