+++ src/backend/controllers/utils/interfaces/index.ts
@@ -1,0 +1,16 @@
+export * from './corpusInterfaces';
+export * from './corpusSuggestionInterfaces';
+export * from './emailInterfaces';
+export * from './exampleInterfaces';
+export * from './exampleSuggestionInterfaces';
+export * from './leaderboardInterfaces';
+export * from './notificationInterfaces';
+export * from './nsibidiCharacterInterfaces';
+export * from './pronunciationInterfaces';
+export * from './requestInterfaces';
+export * from './statInterfaces';
+export * from './suggestionInterfaces';
+export * from './textImageInterfaces';
+export * from './userInterfaces';
+export * from './wordInterfaces';
+export * from './wordSuggestionInterfaces';
